Organize tests for Flowchart Node and Authentication API wrappers into individual files.
